requirements.txt
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
seaborn
requests
scikit-learn
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
tqdm
chromadb
faiss-cpu
pypdf
plotly
st-social-media-links
pillow
supabase
reportlab


